Database methods done
