+++ GameServer/Database/Delete.go
@@ -1,0 +1,1 @@
+package Database
